修改命令； 添加已存在文件判断

+++ bin/index.js
@@ -14,36 +14,41 @@ const spinner = ora('正在下载模板...');
 program.version('1.0.0', '-v, --version')
     .command('create <name>')
     .action((name) => {
-      inquirer.prompt([
-        {
-          name: 'description',
-          message: '请输入项目描述'
-        },
-        {
-          type: 'input',
-          name: 'author',
-          message: '请输入作者名称'
-        }
-      ]).then((answers) => {
-        spinner.start();
-        download('lin-Fore-end/big-project#master', name, {clone: true}, (err) => {
-          if (err) {
-            spinner.fail();
-            console.log(symbols.error, chalk.red('项目创建失败'));
-          } else {
-            spinner.succeed();
-            console.log(symbols.success, chalk.green('项目创建成功'));
-          };
-          const meta = {
-            name,
-            description: answers.description,
-            author: answers.author
-          };
-          const fileName = `${name}/package.json`;
-          const content = fs.readFileSync(fileName).toString();
-          const result = handlebars.compile(content)(meta);
-          fs.writeFileSync(fileName, result);
+      if (fs.existsSync(name)) {
+        // 错误提示项目已存在，避免覆盖原有项目
+        console.log(symbols.error, chalk.red('项目已存在'));
+      } else {
+        inquirer.prompt([
+          {
+            name: 'description',
+            message: '请输入项目描述'
+          },
+          {
+            type: 'input',
+            name: 'author',
+            message: '请输入作者名称'
+          }
+        ]).then((answers) => {
+          spinner.start();
+          download('lin-Fore-end/big-project#master', name, {clone: true}, (err) => {
+            if (err) {
+              spinner.fail();
+              console.log(symbols.error, chalk.red('项目创建失败'));
+            } else {
+              spinner.succeed();
+              console.log(symbols.success, chalk.green('项目创建成功'));
+            };
+            const meta = {
+              name,
+              description: answers.description,
+              author: answers.author
+            };
+            const fileName = `${name}/package.json`;
+            const content = fs.readFileSync(fileName).toString();
+            const result = handlebars.compile(content)(meta);
+            fs.writeFileSync(fileName, result);
+          });
         });
-      });
+      }
     });
 program.parse(process.argv);
